Se corrigen los test, se piso los test del adapter por error
